⌛Refactor: app 모듈 리팩토링 - #1
